+++ src/affliction.js
@@ -3,7 +3,13 @@ Hooks.once("init", () => {
     game.time.advance = async function(seconds, options) {
         const r =  await origin_advance.call(this, seconds, options);
 
-        game.canvas.scene.tokens.forEach(token=>checkAffectionsStage(token, token.actor.itemTypes.affliction))
+        game.canvas.scene.tokens.forEach(token=>{
+            if (token?.actor?.itemTypes?.affliction) {
+                checkAffectionsStage(token, token.actor.itemTypes.affliction)
+            } else {
+                console.log(`Token ${token.name} doesnt have afflictions at all`)
+            }
+        });
 
         return r;
     }